requirements.txt
Django==4.1.4
az-iranian-bank-gateways==1.8.3
django-tinymce==3.5.0
Pillow==9.3.0